Added Tetriminos and beginning logic for gameplay
For now the Ambiq logo will show briefly, followed immediately by the start
of a "game".  At this point the game consists of a single tetrimino (the I shape)
falling down the well until it hits bottom.  It can be moved using the EVK buttons.

The next phase is integration of the game logic to continuosly fetch random tetriminos
as the current on comes to rest.